Moved the checkFront function
Makes sure the cars don't run into each other even on runs when the
controller isn't setting the cars' speeds. It also makes sure that if
the car does need to slow down, that change in speed isn't overwritten
by the next check to try to get the car at the average speed.